requirements.txt
requests==2.18.4
pandas
python-dotenv==0.19.1